Python, macros, operation::dummy, math, serialization changes
- python socket
- python console script
- python plotting updates due to change in serialization
- python testing output folder name fixes
- updated attribute macros
- updated component base math operations
- fixed issue with data_tracker dividing by unit
- macros/attributes header
- updated component serialization
- operation::dummy
- NOINLINE singleton
- minor socket changes
- fix for timem page_rss/virtual_memory

+++ source/python/libpytimemory-component-bundle.hpp
@@ -86,7 +86,7 @@ public:
 
     static void reset()
     {
-        DEBUG_PRINT_HERE("size = %lu", size());
+        DEBUG_PRINT_HERE("size = %lu", (unsigned long) size());
         type::reset();
     }
